Fix callback_tokens_file
